+++ src/crypto/hash/sha256.rs
@@ -59,7 +59,6 @@ mod test {
     }
 
     fn test_nist_vector(filename: &str) {
-        use rustc_serialize::hex::FromHex;
         use std::fs::File;
         use std::io::{BufRead, BufReader};
 
@@ -68,7 +67,7 @@ mod test {
         loop {
             line.clear();
             r.read_line(&mut line).unwrap();
-            if line.len() == 0 {
+            if line.is_empty() {
                 break;
             }
             let starts_with_len = line.starts_with("Len = ");
@@ -76,13 +75,13 @@ mod test {
                 let len: usize = line[6..].trim().parse().unwrap();
                 line.clear();
                 r.read_line(&mut line).unwrap();
-                let rawmsg = line[6..].from_hex().unwrap();
+                let rawmsg = hex::decode(&line[6..].trim()).unwrap();
                 let msg = &rawmsg[..len / 8];
                 line.clear();
                 r.read_line(&mut line).unwrap();
-                let md = line[5..].from_hex().unwrap();
+                let md = hex::decode(&line[5..].trim()).unwrap();
                 let Digest(digest) = hash(msg);
-                assert!(&digest[..] == &md[..]);
+                assert!(digest[..] == md[..]);
             }
         }
     }