+++ src/crypto/secretbox/xsalsa20poly1305.rs
@@ -42,6 +42,10 @@ new_type! {
 pub const MACBYTES: usize = ffi::crypto_secretbox_xsalsa20poly1305_MACBYTES as usize;
 
 /// `gen_key()` randomly generates a secret key
+///
+/// THREAD SAFETY: `gen_key()` is thread-safe provided that you have
+/// called `sodiumoxide::init()` once before using any other function
+/// from sodiumoxide.
 pub fn gen_key() -> Key {
     let mut key = [0; KEYBYTES];
     randombytes_into(&mut key);
@@ -49,6 +53,10 @@ pub fn gen_key() -> Key {
 }
 
 /// `gen_nonce()` randomly generates a nonce
+///
+/// THREAD SAFETY: `gen_key()` is thread-safe provided that you have
+/// called `sodiumoxide::init()` once before using any other function
+/// from sodiumoxide.
 pub fn gen_nonce() -> Nonce {
     let mut nonce = [0; NONCEBYTES];
     randombytes_into(&mut nonce);
@@ -167,7 +175,7 @@ mod test {
             for i in 0..c.len() {
                 c[i] ^= 0x20;
                 // Test the combined mode.
-                assert_eq!(Err(()), open(&mut c, &n, &k));
+                assert_eq!(Err(()), open(&c, &n, &k));
                 // Test the detached mode.
                 let tag = Tag::from_slice(&c[..MACBYTES]).unwrap();
                 assert_eq!(Err(()), open_detached(&mut c[MACBYTES..], &tag, &n, &k));