+++ src/crypto/secretstream/secretstream_macros.rs
@@ -64,7 +64,7 @@ const TAG_FINAL: u8 = $tag_final as u8;
 /// code is generated over all data. A typical encrypted stream simply attaches
 /// `0` as a tag to all messages, except the last one which is tagged as
 /// `Tag::Final`. When decrypting the tag is retrieved and may be used.
-#[derive(Debug, PartialEq)]
+#[derive(Debug, PartialEq, Copy, Clone)]
 pub enum Tag {
     /// Message, the most common tag, that doesn't add any information about the
     /// nature of the message.
@@ -113,6 +113,10 @@ new_type! {
 }
 
 /// `gen_key()` randomly generates a secret key
+///
+/// THREAD SAFETY: `gen_key()` is thread-safe provided that you have
+/// called `sodiumoxide::init()` once before using any other function
+/// from sodiumoxide.
 pub fn gen_key() -> Key {
     let mut key = [0; KEYBYTES];
     randombytes_into(&mut key);
@@ -165,19 +169,22 @@ impl Stream<Push> {
     /// encryption applications can use the password hashing API to get a key
     /// that can be used with the functions below.
     pub fn init_push(key: &Key) -> Result<(Stream<Push>, Header), ()> {
-        let mut header: [u8; HEADERBYTES] = unsafe { mem::uninitialized() };
-        let mut state: $state_name =
-            unsafe { mem::uninitialized() };
+        let mut header = mem::MaybeUninit::<[u8; HEADERBYTES]>::uninit();
+        let mut state = mem::MaybeUninit::uninit();
         let rc = unsafe {
             $init_push_name(
-                &mut state,
-                header.as_mut_ptr(),
+                state.as_mut_ptr(),
+                header.as_mut_ptr() as *mut u8,
                 key.0.as_ptr(),
             )
         };
         if rc != 0 {
             return Err(());
         }
+        // rc == 0 and both state and header are initialized
+        let state = unsafe { state.assume_init() };
+        let header = unsafe { header.assume_init() };
+
         Ok((
             Stream::<Push> {
                 state,
@@ -192,6 +199,26 @@ impl Stream<Push> {
     /// message `m` and its `tag`. Optionally includes additional data `ad`,
     /// which is not encrypted.
     pub fn push(&mut self, m: &[u8], ad: Option<&[u8]>, tag: Tag) -> Result<Vec<u8>, ()> {
+        let buf_len = self.push_check(m, tag)?;
+
+        let mut buf = Vec::with_capacity(buf_len);
+        self.push_impl(m, ad, tag, &mut buf)?;
+        Ok(buf)
+    }
+
+    /// All data (including optional fields) is authenticated. Encrypts a
+    /// message `m` and its `tag`. Optionally includes additional data `ad`,
+    /// which is not encrypted.
+    ///
+    /// The encrypted message is written to the `out` vector, overwriting any existing data there.
+    pub fn push_to_vec(&mut self, m: &[u8], ad: Option<&[u8]>, tag: Tag, out: &mut Vec<u8>) -> Result<(), ()> {
+        let buf_len = self.push_check(m, tag)?;
+        out.clear();
+        out.reserve(buf_len);
+        self.push_impl(m, ad, tag, out)
+    }
+
+    fn push_check(&mut self, m: &[u8], tag: Tag) -> Result<usize, ()> {
         if self.finalized {
             return Err(());
         }
@@ -202,30 +229,31 @@ impl Stream<Push> {
         if tag == Tag::Final {
             self.finalized = true;
         }
+        Ok(m_len + ABYTES)
+    }
 
+    fn push_impl(&mut self, m: &[u8], ad: Option<&[u8]>, tag: Tag, buf: &mut Vec<u8>) -> Result<(), ()> {
         let (ad_p, ad_len) = ad
             .map(|ad| (ad.as_ptr(), ad.len()))
             .unwrap_or((ptr::null(), 0));
-
-        let buf_len = m_len + ABYTES;
-        let mut buf = Vec::with_capacity(buf_len);
-        let rc = unsafe {
-            buf.set_len(buf_len);
-            $push_name(
+        let mut c_len: c_ulonglong = 0;
+        unsafe {
+            let rc = $push_name(
                 &mut self.state,
                 buf.as_mut_ptr(),
-                &mut (buf_len as c_ulonglong),
+                &mut c_len,
                 m.as_ptr(),
-                m_len as c_ulonglong,
+                m.len() as c_ulonglong,
                 ad_p,
                 ad_len as c_ulonglong,
                 tag as u8,
-            )
-        };
-        if rc != 0 {
-            return Err(());
+                );
+            if rc != 0 {
+                return Err(());
+            }
+            buf.set_len(c_len as usize);
         }
-        Ok(buf)
+        Ok(())
     }
 
     /// Create a ciphertext for an empty message with the `TAG_FINAL` added
@@ -234,6 +262,7 @@ impl Stream<Push> {
     pub fn finalize(mut self, ad: Option<&[u8]>) -> Result<Vec<u8>, ()> {
         self.push(&[], ad, Tag::Final)
     }
+
 }
 
 impl Stream<Pull> {
@@ -241,11 +270,10 @@ impl Stream<Pull> {
     /// will not be required any more for subsequent operations. `Err(())` is
     /// returned if the header is invalid.
     pub fn init_pull(header: &Header, key: &Key) -> Result<Stream<Pull>, ()> {
-        let mut state: $state_name =
-            unsafe { mem::uninitialized() };
+        let mut state = mem::MaybeUninit::uninit();
         let rc = unsafe {
             $init_pull_name(
-                &mut state,
+                state.as_mut_ptr(),
                 header.0.as_ptr(),
                 key.0.as_ptr(),
             )
@@ -259,6 +287,9 @@ impl Stream<Pull> {
         } else if rc != 0 {
             return Err(());
         }
+        // rc == 0 and state is initialized
+        let state = unsafe { state.assume_init() };
+
         Ok(Stream::<Pull> {
             state: state,
             finalized: false,
@@ -278,6 +309,33 @@ impl Stream<Pull> {
     /// Applications will typically use a `while stream.is_not_finalized()`
     /// loop to authenticate and decrypt a stream of messages.
     pub fn pull(&mut self, c: &[u8], ad: Option<&[u8]>) -> Result<(Vec<u8>, Tag), ()> {
+        let m_len = self.pull_check(c)?;
+        let mut buf = Vec::with_capacity(m_len);
+        let tag = self.pull_impl(c, ad, &mut buf)?;
+        Ok((buf, tag))
+    }
+
+    /// Verifies that `c` is a valid ciphertext with a correct authentication tag
+    /// given the internal state of the `Stream` (ciphertext streams cannot be
+    /// decrypted out of order for this reason). Also may validate the optional
+    /// unencrypted additional data `ad` using the authentication tag attached to
+    /// `c`. Finally decrypts the ciphertext and tag, and checks the tag
+    /// validity.
+    /// If any authentication fails, the stream has already been finalized, or if
+    /// the tag byte for some reason does not correspond to a valid `Tag`,
+    /// returns `Err(())`. Otherwise returns the plaintext and the tag.
+    /// Applications will typically use a `while stream.is_not_finalized()`
+    /// loop to authenticate and decrypt a stream of messages.
+    ///
+    /// The decrypted message is written to the `out` vector, overwriting any existing data there.
+    pub fn pull_to_vec(&mut self, c: &[u8], ad: Option<&[u8]>, out: &mut Vec<u8>) -> Result<Tag, ()> {
+        let m_len = self.pull_check(c)?;
+        out.clear();
+        out.reserve(m_len);
+        self.pull_impl(c, ad, out)
+    }
+
+    fn pull_check(&self, c: &[u8]) -> Result<usize, ()> {
         if self.finalized {
             return Err(());
         }
@@ -290,32 +348,38 @@ impl Stream<Pull> {
         if m_len > messagebytes_max() {
             return Err(());
         }
+        Ok(m_len)
+    }
+
+    fn pull_impl(&mut self, c: &[u8], ad: Option<&[u8]>, buf: &mut Vec<u8>) -> Result<Tag, ()> {
+        let mut tag: u8 = 0;
+        let mut m_len: c_ulonglong = 0;
         let (ad_p, ad_len) = ad
             .map(|ad| (ad.as_ptr(), ad.len()))
             .unwrap_or((ptr::null(), 0));
-        let mut tag: u8 = unsafe { mem::uninitialized() };
-        let mut buf = Vec::with_capacity(m_len);
-        let rc = unsafe {
-            buf.set_len(m_len);
-            $pull_name(
+        unsafe {
+            let rc = $pull_name(
                 &mut self.state,
                 buf.as_mut_ptr(),
-                &mut (m_len as c_ulonglong),
+                &mut m_len,
                 &mut tag,
                 c.as_ptr(),
-                c_len as c_ulonglong,
+                c.len() as c_ulonglong,
                 ad_p,
                 ad_len as c_ulonglong,
-            )
-        };
-        if rc != 0 {
-            return Err(());
+                );
+            if rc != 0 {
+                return Err(());
+            }
+            // rc == 0 and tag is initialized
+            buf.set_len(m_len as usize);
         }
+
         let tag = Tag::from_u8(tag)?;
         if tag == Tag::Final {
             self.finalized = true;
         }
-        Ok((buf, tag))
+        Ok(tag)
     }
 }