+++ src/lib.rs
@@ -58,14 +58,13 @@
 #![warn(unused_qualifications)]
 #![cfg_attr(not(feature = "std"), no_std)]
 #![cfg_attr(not(feature = "std"), feature(alloc))]
+#![deny(clippy::all)]
 
 extern crate libsodium_sys as ffi;
 
-#[macro_use]
-extern crate ctor;
-extern crate libc;
 #[cfg(test)]
-extern crate rustc_serialize;
+extern crate hex;
+extern crate libc;
 #[cfg(any(test, feature = "serde"))]
 extern crate serde;
 #[cfg(not(feature = "std"))]
@@ -91,22 +90,11 @@ mod prelude {
 /// thread-safe
 ///
 /// `init()` returns `Ok` if initialization succeeded and `Err` if it failed.
-#[deprecated(
-    since = "0.2.2",
-    note = "libsodium is automatically initialized by sodiumoxide now."
-)]
 pub fn init() -> Result<(), ()> {
     if unsafe { ffi::sodium_init() } >= 0 {
         Ok(())
     } else {
         Err(())
-    }
-}
-
-#[ctor]
-fn init_real() {
-    unsafe {
-        ffi::sodium_init();
     }
 }